Updated docs to add clarification about compile dependencies, ports and added a diagram from the forums.
NOTE: This is my first attempt to do something like this.  I know the image is remote but I don't know how to add it to the server.  Also am new to markdown.

Accept suggestions.  If this works out as I learn more about the product I will update pages.