Fix latency spike during connection failures
    The issue was when a connection is not there, because a peer node is
down, we were queuing up the messages till they would timeout and fail
eventually. This would lead to higher overall average latency and much
higher 99th latencies. The solution is to not queue up messages and fail
the requests till here is a good connection
    o Drop the messages and respond with error immediately when a
connection is not available
    o Do the same when the connection is in a connecting state
    o Split the dnode_peer_conn into 2 functions to get the peer server
and then to get a connection to it
    o Mark the node as down after some failure attempts and keep it down
for 30 secs default.