Fix return type for all forward request apis
Today we do not return any error if the request forwarding fails for any
reason. So there is no way caller can know what exactly happened. This
is an important patch which tries to add the notion of api failures and
react accordingly. This could open enough avenues like failing over to
another node for remote dc replication, or simply maintaining list of
messages that failed to replicate etc.