Add a counter to tailq
Whereever we are using this tailq, we are keeping a counter seperately.
For example. free messages, connection inq and outq etc. This method is
mainly error prone. Added a counter to the tailq header itself, so that
any change in the tailq list will update the counter. It is imperative
to use the given macros to update the linked list.
Also added test cases to the dynomite-test