Fix crash during node replacement
When a node gets replaced, we destroy the connection pool but we were
cancelling the scheduled task and then recreating it as connection
errored callback, leading to a orphaned scheduled task around. This
would lead to accessing a connection pool object that can be freed.
The solution is to cancel the task just before freeing the object.