Avoid accessing 'ctx->stats' if it was never created
If a call to 'stats_create()' fails from 'core_stats_create()', we
still access the 'stats' structure on the teardown path in multiple
places. This causes Dynomite to crash in multiple places.

This patch fixes the crashes by avoiding accessing the invalid memory
if it wasn't allocated in the first place.