Avoid implicit interpretation of return type for dyn_aes_encrypt_msg()
dyn_aes_encrypt_msg() returns a type of rstatus_t, however the function
returns a status only on an error and returns a 'count' on success
which the caller implicitly interprets as '# of bytes'. This pattern
is error prone and could lead to hard to track down bugs.

This patch also returns the right error code on the OOM case and also
adds checking on the caller side for this new error code DN_ENOMEM.

This was noticed while debugging a core dump. While this patch is only
for clean up, a future patch will change some more code in this area
to fix a crash bug.