Consistency for SMEMBERS when using DC_SAFE_QUORUM
Although the data written to the same set in different replics will be
the same, the order in which they're returned in each replica might
be different. This causes the checksums of each replica set to be
different, causing SMEMBERS to fail under DC_SAFE_QUORUM.

This patch addresses this by rewriting all SMEMBERS calls to the
following (only when DC_SAFE_QUORUM is set):
 * SMEMBERS <set> -> SORT <set> ALPHA

This ensures that the returned set is in the same order across replicas
and hence ensures that their checksums match.

A mini-framework for rewriting queries is introduced to make future
query rewrites easier. Note that this is different from query
fragmenting which already exists today for queries like MGET.