Error responses should contain mbufs with appropriate error strings
We've noticed that when the outstanding message queue for a connection
(conn->omsg_q) reaches the MAX_CONN_QUEUE_SIZE, we set the
ENOTRECOVERABLE state. This eventually leads a server to call
server_ack_err() which sends a error response message (struct msg*)
back to the client. This message however does not contain a mbuf
with an error string, causing failures in the client.

Eg: The dyn_dnode_client tries to encrypt this empty message in
dnode_rsp_send_next(), which returns an error and eventually
leads to a crash later on.

This patch fixes this by filling in these error responses with
a valid mbuf which contains an error string.

No further cleanup is done here, however, some clean up needs to
be done around this code in the future.