Fix omission of Debug flags via #include in dyn_queue.h
We ran into a very strange problem recently with full-debug mode
enabled. On any call to _conn_get(), we expected to get a 'struct conn'
object completely initialized with default values (mostly 0 & NULL).

However, immediately after returning from _conn_get(), certain members
of the object got completely corrupted after the 'retq' instruction
was executed on returning from _conn_get(). This was very confusing
as the retq instruction has no power to overwrite members of that
address which lived in the heap, and there was no overflow of the
stack as well.

Upon further investigation it was found that the size of the
'struct conn' object was completely different in _conn_get() and
in other parts of the code.

sizeof(struct conn) in _conn_get() : 544 bytes
sizeof(struct conn) in other files: 672 bytes

This led me to investigate all individual members of 'struct conn'.
My finding was that any member initialized by 'TAILQ_ENTRY(xyz)'
was basically a struct with a nested debug struct:

 #define TAILQ_ENTRY(type)                                          \
   struct {                                                         \
     struct type *tqe_next;  /* next element */                     \
     struct type **tqe_prev; /* address of previous next element */ \
     TRACEBUF                                                       \
   }

'TRACEBUF' is a struct that's only compiled in full-debug mode. So what
this means is that _conn_get() was interpreting 'struct conn' as the
release version of the object and the rest of the code was interpreting
'struct conn' as the debug version of the object. Hence, _conn_get()
would only allocate memory for the release version of 'struct conn'
which was much smaller than required, and the rest of the code expected
the debug version, hence overwriting other members of the 'struct conn'
returned by _conn_get().

The reason that _conn_get() was interpreting it as the release-version
even though full-debug mode was on, is that after the following commit:
https://github.com/Netflix/dynomite/commit/192af493431208e35f1f673780d29d672bff5072

, the dyn_queue.h file (which contains the definition of 'TRACEBUF'),
lost the connection to dyn_core.h which contained the initializations
of the Debug flags.

To fix this, I moved the debug flag definitions to dyn_types.h, since
we're treating that as the bare-minimum header required for most of the
files (since it doesn't depend on any other dyn_*** file), and included
that in dyn_queue.h. This now allows dyn_connection_internal.h/c to
be aware of Debug mode if it's enabled, hence fixing the above issue.