Allow multi-buffer payload decryption
In many cases, queries or responses that spanned multiple mbufs
would not get decrypted correctly causing cross-DC traffic to be
dropped silently which resulted in a failure to replicate data in
different DCs.

After a lot of debugging, this was found to be due to a typo
introduced in a previous commit:
https://github.com/Netflix/dynomite/commit/91e8fe38ea66b79b5241f02169fa04e2ad9a50ea

This typo is now fixed and multi-buffer decryption works seamlessly.