[WIP] Read repairs for primary keys
This patch adds read repairs for primary keys. It specifically works
only with the following commands at the moment:
 - SET
 - GET
 - DEL
 - APPEND

Repiars need to be enabled under a flag in the YAML configuration. It is
enabled only if 'repairs_enabled: true' is specified.

This current implementation of repairs is not supported and returns the
values in a custom format:
1) E/N (exists or not-exists)
2) Value itself
3) Timestamp of value

This custom format will be removed in future patches to standardize on
the regular Redis wire protocol.

Currently the timestamp is generated on the server side. Client side
timestamp support will be added in future patches.

TODO:-
 - Add client side timestamp support.
 - Add repair support for secondary keys.
 - Remove custom result format.
 - Add support for tracking expired keys.
 - Add background repir process.
 - Add documentation on repair implementation.