Allow multi-mbuf requests to work cleanly
This patch allows multi-mbuf requests that are crafted (as opposed to
being received from the network) to be parsed successfully.

The reason the behavior is different between the crafted case and the
recieved from network case is that when a multi-mbuf payload is being
received from the network, we only get one mbuf at a time which we
parse before receiving the next mbuf and parsing it. Due to this, the
parser makes assumptions about which mbufs it has already parsed; or in
other words, it always picks the tail of the 'msg->mhdr' list and tries
to parse it and since we always parse every mbuf before receiving the
next, we'll parse every mbuf.

In the crafting case, we have all the mbufs at once, so parsing from the
tail would be incorrect since we would skip all the mbufs before that.

This patch also modifies msg_prepend_format() since it makes an
assumption that we will write only to a single mbuf. This patch changes
that by allowing to prepend multiple mbufs to a single 'msg' in one
call which is required for rewriting scripts with potentially large
arguments.