Allow toggling of read repairs via REST
This patch moves the ultimate switch for read repairs from
'struct context' to a global variable called 'g_read_repairs_enabled'.

This is so that we can toggle read repairs on or off via a REST call.
Having it on and off on different nodes in the cluster will not cause
a problem as only the coordinating node for a query will decide whether
that query is written with metadata and/or a read query does a repair.

The way to toggle repairs is by making the following REST call:

   curl http://localhost:22222/toggle_read_repairs