Return non-bulk reply for single key look ups w/ read repairs
The adjust_rsp_buffers_for_client() function is meant to remove the
metadata that dynomite receives from Redis before returning only the
requested value(s). This was always returned as a multi-bulk response
which can confuse some clients that have requested only a single key.

This patch changes that to return exactly what the client expects.