ROUTING_ALL_NODES_ALL_RACKS_ALL_DCS routing option shouldn't enforce quorum
The ROUTING_ALL_NODES_ALL_RACKS_ALL_DCS option forwards a message to
each and every node in the cluster. This is only used for commands that
don't have any sharding strategy (like SCRIPT LOAD).

Enforcing quorum on these types of commands do not make any sense, so
this patch overrides the consistency level for such queries to DC_ONE
regardless of the configuration.