Add option to force swallow for messages forwarded to non-replica peers
Under ROUTING_ALL_NODES_ALL_RACKS_ALL_DCS, we forward the message to
every node in the cluster and hence we don't want to hear back
responses from nodes that are non-replicas.

This patch adds an option to req_forward_to_peer() to force swallowing
of all responses (except the one that goes to the local datastore).