Treat each rack continuum as an array
Each rack has a 'struct continuum' pointer in it. This is used to determine
which token belongs to which node (see dyn_vnode.c).

While it has only a single pointer, the code uses it as though it's
an array of struct pointers. This potentially could lead to a lot of
confusion.

This patch addresses this by explicitly changing the member to an array
of structs.

Manually tested to make sure that there is no change to how the keys are
sharded.