Allow commands with optional arguments to be read-repairable and support ZSET
The read repair infra did not allow for commands with optional support to
be read-repairable. Now, this functionality is added which lets us support
more Redis commands.

ZADD is the first command with optional arguments to have read-repair support.
ZSCORE and ZREM are also supported for read repairs.

As with the other commands, range queries are not repairable yet, and will be
added as a follow on patch.