Fix crash during a MSG_PARSE_AGAIN parse state w/ repairs
When read repairs are enabled, we attempt to keep a track of the last
parsed mbuf if the request is large enough to be split across multiple
buffers.

The code would previously update the latest_parsed_mbuf_idx field if the
current token length surpassed (mbuf)->last. This turns out to be incorrect
since we may parse with partial network reads as well.

The fix is to check if mbuf_full() before updating said index. Confirmed its
working with large requests across mbufs and partial network read requests.