Auto cleanup repair Metadata on deletes if all replicas agree
This patch implements the automatic cleanup of metadata for delete
commands (DEL, ZREM, HDEL and SREM) when ALL the replicas in the
quorum set agree.

This is to ensure that we don't have "tombstones" present indefinitely
for deleted keys and fields.

For commands that don't receive all replica responses, the "rem set"
data will still be present. However, this would be a minority of the
cases, and we can implement logic in the side car to get rid of this.