Support the Redis "QUIT" command properly
The QUIT command so far was handled by just closing the connection on
the server side and did not respect the contract of the QUIT command.

The Redis documentation states that the QUIT command must return
"+OK\r\n" back to the client so that the client can close the connection
on its side.

This patch takes advantage of simulating a datastore response from the
previous patch to achieve this.