Add ownership information for dnode_peer/server errors and fix cross-DC checksum logic
When we hit an error trying to connect to a dnode peer or a server,
we create an appropriate error message but don't link the respective
connection to the error response 'msg'. This is necessary at least for
processing DC_EACH_SAFE_QUORUM responses.

Also, checksums for cross-DC messages were not done properly since
msg_payload_crc32() only accounted for intra-DC messages. More explained
in code comments.