Replicas shouldn't update top level rem-set on final field delete
This applies to hashmaps, sorted sets and sets.

Consider the following example:
If a sorted set in Dynomite has 2 or more items across replicas,
but if one of the replicas has a missed update(s) and therefore has
only 1 item, removing that item using dynomite previously updated
the top level REM set in that replica.

This is incorrect because now trying to read the other (non-deleted)
field, would cause read repairs to compare its timestamp against the
TS in the top level REM set in the inconsistent replica, and delete
the item incorrectly thinking that someone removed the entire map/set/set
in that broken replica.

This patch fixes it by not updating the top level REM set for field
deletes. Only a DEL will update the top level REM set for maps/sets/zsets.

This was noticed in a TEST cluster.